// src/types/database.types.ts
export interface Database {
  public: {
    Tables: {
// ...
      media_files: {
        Row: {
          created_at: string
          description: string | null
          file_path: string
          id: string
          keywords: string[] | null
          name: string
          url: string
          user_id: string
        }
        Insert: {
          created_at?: string
          description?: string | null
          file_path: string
          id?: string
          keywords?: string[] | null
          name: string
          url: string
          user_id: string
        }
        Update: {
          created_at?: string
          description?: string | null
          file_path?: string
          id?: string
          keywords?: string[] | null
          name?: string
          url?: string
          user_id?: string
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      proofing_status: PublicEnumProofingStatus
      task_priority: PublicEnumTaskPriority
      task_status: PublicEnumTaskStatus
      team_role: PublicEnumTeamRole
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

export type Tables<T extends keyof Database["public"]["Tables"]> =
  Database["public"]["Tables"][T]["Row"]
export type TablesInsert<T extends keyof Database["public"]["Tables"]> =
  Database["public"]["Tables"][T]["Insert"]
export type TablesUpdate<T extends keyof Database["public"]["Tables"]> =
  Database["public"]["Tables"][T]["Update"]